# Tes Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Epresence API Documentation",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Get Authorization Token",
                "description": "Enter your credential here",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@email.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return authorization token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/karyawan": {
            "get": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Get List Karyawan",
                "operationId": "getKaryawan",
                "responses": {
                    "default": {
                        "description": "Return list of karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/karyawan/presences/{user}": {
            "get": {
                "tags": [
                    "Karyawan"
                ],
                "summary": "Get List presences of karyawan",
                "operationId": "getKaryawanPresences",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID User",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Return list presences of karyawan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/presence": {
            "get": {
                "tags": [
                    "Presence"
                ],
                "summary": "Get user Presences list",
                "operationId": "getPresences",
                "responses": {
                    "default": {
                        "description": "Return list of user presences",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Presence"
                ],
                "summary": "Create presence data",
                "description": "Input type and time presence here",
                "operationId": "createPresence",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "IN",
                                            "OUT"
                                        ]
                                    },
                                    "waktu": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "type": "IN",
                                    "waktu": "2024-04-06 08:00:00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return authorization token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/presence/{presence}": {
            "put": {
                "tags": [
                    "Presence"
                ],
                "summary": "Approve presence",
                "description": "Input presence id here",
                "operationId": "approvePresence",
                "parameters": [
                    {
                        "name": "presence",
                        "in": "path",
                        "description": "ID presence",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Return approved presence data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}